fix: 빌드 에러 해결 및 아바타 정보 DTO에 맞게 수정
