+++ src/types/productDetail.ts
@@ -17,3 +17,16 @@ export type ProductDetailResponse = {
   liked: boolean;
   variant: ProductVariant[];
 };
+
+export const initialProductDetail: ProductDetailResponse = {
+  id: 1,
+  productName: "",
+  brand: "",
+  price: 0,
+  sale: 0,
+  content: "",
+  images: [],
+  wishlistCount: 0,
+  liked: false,
+  variant: [],
+};
